# platform/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TerraGenius AI API",
    "description": "API for generating and managing infrastructure-as-code.\n\n## Authentication Flow\n\n1. User signs up or logs in using Firebase Authentication (email/password, Google, etc.) via the frontend.\n2. Frontend obtains a Firebase ID token.\n3. Frontend POSTs the ID token to `/auth/token`.\n4. Backend verifies the Firebase token and returns a custom JWT (`accessToken`).\n5. All protected API calls require `Authorization: Bearer <accessToken>`.\n6. The backend validates the JWT and attaches user context.\n\n**Public endpoints:** `/auth/token`, `/status`, `/health`, `/signup`, `/login`, `/reset-password`\n**Protected endpoints:** All `/api/` routes, `/generate`, `/deployments`, `/billing`, `/audit-logs`, etc.\n\n**Future Ready:** To swap auth providers (e.g. Auth0, Cognito), only the `/auth/token` logic and token validation need to change. The rest of the API and security model remains the same.\n\n**Testing:** Use the provided `test_auth_flow.js` script to verify the full login → token exchange → protected API flow.\n\n**Security:** Never send Firebase tokens directly to protected endpoints. Always use the custom JWT returned from `/auth/token`.\n",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Mock API Server"
    }
  ],
  "paths": {
...
    "/deployments/{projectId}": {
      "get": {
        "summary": "List deployments for a project",
        "description": "Retrieve all deployment records for a specific project",
        "tags": ["deployments"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Project identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "List of deployments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new deployment record",
        "description": "Create a new deployment record in Firestore",
        "tags": ["deployments"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Project identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deployment record created successfully"
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/token endpoint"
      }
    },
    "schemas": {
...
      "ValidationStepLog": {
        "type": "object",
        "properties": {
          "stage": {"type": "string", "example": "generate"},
          "status": {"type": "string", "example": "success"},
          "message": {"type": "string", "example": "AI response received."},
          "timestamp": {"type": "string", "format": "date-time"}
        },
        "required": ["stage", "status", "message", "timestamp"]
      },
...
      "DeploymentRequest": {
        "type": "object",
        "properties": {
          "projectId": { "type": "string" },
          "generationId": { "type": "string" },
          "provider": { "type": "string" },
          "region": { "type": "string" },
          "credentialsId": { "type": "string" }
        },
        "required": ["projectId", "generationId", "provider", "region", "credentialsId"]
      },
      "DeploymentListResponse": {
        "type": "object",
        "properties": {
          "deployments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeploymentResponse"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": ["deployments", "total"]
      },
      "DeploymentResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "projectId": { "type": "string" },
          "generationId": { "type": "string" },
          "provider": { "type": "string" },
          "region": { "type": "string" },
          "credentialsId": { "type": "string" },
          "status": { "type": "string" },
          "logs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationStepLog" }
          },
          "outputs": {
            "type": "object"
          },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        },
        "required": ["id", "userId", "projectId", "generationId", "provider", "region", "credentialsId", "status", "logs", "outputs", "createdAt", "updatedAt"]
      },
...
    }
  }
}